Break up the rewrite module
